Put more logic in the pure FSS module, improve module imports
